test: continue scanner test decomposition
