+++ tests/test_device_scanner.py
@@ -16,7 +16,6 @@ from custom_components.thessla_green_modbus.modbus_exceptions import (
 from custom_components.thessla_green_modbus.registers.loader import get_registers_by_function
 from custom_components.thessla_green_modbus.scanner.core import (
     DeviceCapabilities,
-    ScannerDeviceInfo,
     ThesslaGreenDeviceScanner,
 )
 from custom_components.thessla_green_modbus.scanner_helpers import _format_register_value
@@ -193,191 +192,7 @@ async def test_scan_device_connection_failure():
         await scanner.close()
 
 
-async def test_scan_device_firmware_unavailable(caplog):
-    """Missing firmware registers should log info and report unknown firmware."""
-    empty_regs = {4: {}, 3: {}, 1: {}, 2: {}}
-    with patch.object(
-        ThesslaGreenDeviceScanner, "_load_registers", AsyncMock(return_value=(empty_regs, {}))
-    ):
-        scanner = await ThesslaGreenDeviceScanner.create("192.168.1.1", 502, 10)
 
-    async def fake_read_input(*args, skip_cache=False):
-        if len(args) == 2:
-            address, count = args
-        else:
-            _, address, count = args
-        if address == 0 and count >= 16:
-            return None
-        if count == 1 and address in (
-            INPUT_REGISTERS["version_major"],
-            INPUT_REGISTERS["version_minor"],
-            INPUT_REGISTERS["version_patch"],
-        ):
-            return None
-        return [1] * count
-
-    async def fake_read_holding(*args, **kwargs):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [1] * count
-
-    async def fake_read_coil(*args):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [False] * count
-
-    async def fake_read_discrete(*args):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [False] * count
-
-    with patch("pymodbus.client.AsyncModbusTcpClient") as mock_client_class:
-        mock_client = AsyncMock()
-        mock_client.connect.return_value = True
-        mock_client_class.return_value = mock_client
-
-        with (
-            patch.object(scanner, "_read_input", AsyncMock(side_effect=fake_read_input)),
-            patch.object(scanner, "_read_holding", AsyncMock(side_effect=fake_read_holding)),
-            patch.object(scanner, "_read_coil", AsyncMock(side_effect=fake_read_coil)),
-            patch.object(scanner, "_read_discrete", AsyncMock(side_effect=fake_read_discrete)),
-        ):
-            caplog.set_level(logging.WARNING)
-            scanner.connection_mode = CONNECTION_MODE_TCP
-            result = await scanner.scan_device()
-
-    assert result["device_info"]["firmware"] == "Unknown"
-    assert "Failed to read firmware version registers" in caplog.text
-
-
-async def test_scan_device_firmware_bulk_fallback():
-    """Bulk firmware read failure should fall back to individual reads."""
-    empty_regs = {4: {}, 3: {}, 1: {}, 2: {}}
-    with patch.object(
-        ThesslaGreenDeviceScanner, "_load_registers", AsyncMock(return_value=(empty_regs, {}))
-    ):
-        scanner = await ThesslaGreenDeviceScanner.create("192.168.1.1", 502, 10)
-
-    async def fake_read_input(*args, skip_cache=False):
-        if len(args) == 2:
-            address, count = args
-        else:
-            _, address, count = args
-        if address == 0 and count >= 16:
-            return None
-        if count == 1 and address == INPUT_REGISTERS["version_major"]:
-            return [4]
-        if count == 1 and address == INPUT_REGISTERS["version_minor"]:
-            return [85]
-        if count == 1 and address == INPUT_REGISTERS["version_patch"]:
-            return [0]
-        return [1] * count
-
-    async def fake_read_holding(*args, **kwargs):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [1] * count
-
-    async def fake_read_coil(*args):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [False] * count
-
-    async def fake_read_discrete(*args):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [False] * count
-
-    with patch("pymodbus.client.AsyncModbusTcpClient") as mock_client_class:
-        mock_client = AsyncMock()
-        mock_client.connect.return_value = True
-        mock_client_class.return_value = mock_client
-
-        with (
-            patch.object(scanner, "_read_input", AsyncMock(side_effect=fake_read_input)),
-            patch.object(scanner, "_read_holding", AsyncMock(side_effect=fake_read_holding)),
-            patch.object(scanner, "_read_coil", AsyncMock(side_effect=fake_read_coil)),
-            patch.object(scanner, "_read_discrete", AsyncMock(side_effect=fake_read_discrete)),
-        ):
-            scanner.connection_mode = CONNECTION_MODE_TCP
-            result = await scanner.scan_device()
-
-    assert result["device_info"]["firmware"] == "4.85.0"
-
-
-async def test_scan_device_firmware_partial_bulk_fallback():
-    """Partial firmware bulk read should fall back to individual reads."""
-    empty_regs = {4: {}, 3: {}, 1: {}, 2: {}}
-    with patch.object(
-        ThesslaGreenDeviceScanner, "_load_registers", AsyncMock(return_value=(empty_regs, {}))
-    ):
-        scanner = await ThesslaGreenDeviceScanner.create("192.168.1.1", 502, 10)
-
-    async def fake_read_input(*args, skip_cache=False):
-        if len(args) == 2:
-            address, count = args
-        else:
-            _, address, count = args
-        if address == 0 and count >= 16:
-            return [4, 85]
-        if address >= 16:
-            return []
-        if count == 1 and address == INPUT_REGISTERS["version_patch"]:
-            return [0]
-        if count == 1 and address == INPUT_REGISTERS["version_major"]:
-            return [4]
-        if count == 1 and address == INPUT_REGISTERS["version_minor"]:
-            return [85]
-        return [1] * count
-
-    async def fake_read_holding(*args, **kwargs):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [1] * count
-
-    async def fake_read_coil(*args):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [False] * count
-
-    async def fake_read_discrete(*args):
-        if len(args) == 2:
-            _, count = args
-        else:
-            _, _, count = args
-        return [False] * count
-
-    with patch("pymodbus.client.AsyncModbusTcpClient") as mock_client_class:
-        mock_client = AsyncMock()
-        mock_client.connect.return_value = True
-        mock_client_class.return_value = mock_client
-
-        with (
-            patch.object(scanner, "_read_input", AsyncMock(side_effect=fake_read_input)),
-            patch.object(scanner, "_read_holding", AsyncMock(side_effect=fake_read_holding)),
-            patch.object(scanner, "_read_coil", AsyncMock(side_effect=fake_read_coil)),
-            patch.object(scanner, "_read_discrete", AsyncMock(side_effect=fake_read_discrete)),
-        ):
-            scanner.connection_mode = CONNECTION_MODE_TCP
-            result = await scanner.scan_device()
-
-    assert result["device_info"]["firmware"] == "4.85.0"
 
 
 async def test_scan_blocks_propagated():
@@ -832,153 +647,6 @@ async def test_temperature_unavailable_no_warning(caplog):
     assert scanner._is_valid_register_value("outside_temperature", SENSOR_UNAVAILABLE) is True
     assert "outside_temperature" not in caplog.text
 
-
-async def test_capabilities_detect_schedule_keywords():
-    """Ensure capability detection considers scheduling related registers."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    scanner.available_registers["holding_registers"].add("airing_start_time")
-    caps = scanner._analyze_capabilities()
-    assert caps.weekly_schedule is True
-
-
-@pytest.mark.parametrize(
-    "register",
-    ["constant_flow_active", "supply_air_flow", "supply_flow_rate", "cf_version"],
-)
-async def test_constant_flow_detected_from_various_registers(register):
-    """Constant flow capability is detected from different register names."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    scanner.available_registers = {
-        "input_registers": {register},
-        "holding_registers": set(),
-        "coil_registers": set(),
-        "discrete_inputs": set(),
-    }
-    caps = scanner._analyze_capabilities()
-    assert caps.constant_flow is True
-
-
-async def test_analyze_capabilities():
-    """Test capability analysis."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.1.100", 502, 10)
-
-    # Mock available registers
-    scanner.available_registers = {
-        "input_registers": {"constant_flow_active", "outside_temperature"},
-        "holding_registers": {"gwc_mode", "bypass_mode"},
-        "coil_registers": {"power_supply_fans"},
-        "discrete_inputs": {"expansion"},
-    }
-
-    capabilities = scanner._analyze_capabilities()
-
-    assert capabilities.constant_flow is True
-    assert capabilities.gwc_system is True
-    assert capabilities.bypass_system is True
-    assert capabilities.expansion_module is True
-    assert capabilities.sensor_outside_temperature is True
-
-
-async def test_analyze_capabilities_flag_presence():
-    """Capabilities should reflect register presence and absence."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.1.100", 502, 10)
-
-    # Positive case: registers exist
-    scanner.available_registers = {
-        "input_registers": {"constant_flow_active", "outside_temperature"},
-        "holding_registers": {"gwc_mode", "airing_start_time"},
-        "coil_registers": set(),
-        "discrete_inputs": {"expansion"},
-    }
-    capabilities = scanner._analyze_capabilities()
-
-    assert capabilities.constant_flow is True
-    assert capabilities.sensor_outside_temperature is True
-    assert capabilities.expansion_module is True
-    assert capabilities.gwc_system is True
-    assert capabilities.weekly_schedule is True
-
-    # Negative case: registers absent
-    scanner.available_registers = {
-        "input_registers": set(),
-        "holding_registers": set(),
-        "coil_registers": set(),
-        "discrete_inputs": set(),
-    }
-    capabilities = scanner._analyze_capabilities()
-
-    assert capabilities.constant_flow is False
-    assert capabilities.sensor_outside_temperature is False
-    assert capabilities.expansion_module is False
-    assert capabilities.gwc_system is False
-    assert capabilities.weekly_schedule is False
-
-
-async def test_capability_rules_detect_heating_and_bypass():
-    """Capability rules infer heating and bypass systems from registers."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.1.100", 502, 10)
-    scanner.available_registers = {
-        "input_registers": {"heater_active"},
-        "holding_registers": {"bypass_position"},
-        "coil_registers": set(),
-        "discrete_inputs": set(),
-    }
-
-    capabilities = scanner._analyze_capabilities()
-
-    assert capabilities.heating_system is True
-    assert capabilities.bypass_system is True
-
-
-async def test_scan_device_includes_capabilities_in_device_info():
-    """Detected capabilities are exposed on device info returned by scanner."""
-    scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-    info = ScannerDeviceInfo(
-        model="m", firmware="f", serial_number="s", capabilities=["heating_system"]
-    )
-    caps = DeviceCapabilities(heating_system=True)
-
-    with patch.object(scanner, "scan", AsyncMock(return_value=(info, caps, {}))):
-        scanner.connection_mode = CONNECTION_MODE_TCP
-        result = await scanner.scan_device()
-
-    assert result["device_info"]["capabilities"] == ["heating_system"]
-
-
-async def test_capability_count_includes_booleans(caplog):
-    """Log should count boolean capabilities even though bool is an int subclass."""
-    empty_regs = {4: {}, 3: {}, 1: {}, 2: {}}
-    with patch.object(
-        ThesslaGreenDeviceScanner,
-        "_load_registers",
-        AsyncMock(return_value=(empty_regs, {})),
-    ):
-        scanner = await ThesslaGreenDeviceScanner.create("host", 502, 10)
-
-    caps = DeviceCapabilities(
-        basic_control=True,
-        expansion_module=True,
-        temperature_sensors={"outside"},
-        temperature_sensors_count=1,
-    )
-
-    with patch.object(scanner, "_analyze_capabilities", return_value=caps):
-        with patch("pymodbus.client.AsyncModbusTcpClient") as mock_client_class:
-            mock_client = AsyncMock()
-            mock_client.connect.return_value = True
-            mock_client_class.return_value = mock_client
-
-            with (
-                patch.object(scanner, "_read_input", AsyncMock(return_value=[])),
-                patch.object(scanner, "_read_holding", AsyncMock(return_value=[])),
-                patch.object(scanner, "_read_coil", AsyncMock(return_value=[])),
-                patch.object(scanner, "_read_discrete", AsyncMock(return_value=[])),
-            ):
-                with caplog.at_level(logging.INFO):
-                    scanner.connection_mode = CONNECTION_MODE_TCP
-                    await scanner.scan_device()
-
-    assert any("2 capabilities" in record.message for record in caplog.records)
 
 
 async def test_scan_populates_device_name():
